examples/integration/gemini-live-adapter/requirements.txt
google-genai>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
